Convert commitsDelta to typescript
Also extract the API response for the Commits API
to a runtime type and use in the action.
