+++ client/deltas/index.ts
@@ -3,6 +3,7 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 import { observeDelta } from 'brookjs';
 
 export * from './authorDelta';
+export * from './commitsDelta';
 
 export const applyDelta = (...args: Array<any>) =>
   composeWithDevTools(applyMiddleware(observeDelta(...args)));
